file cleanup and comments
